LayoutTests/fast/repaint/content-visibility-hidden-container-with-abs-pos-c-v-auto-child-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('50 50') is not -1
PASS successfullyParsed is true

TEST COMPLETE
